+++ lib/utils/i_validator.dart
@@ -1,3 +1,0 @@
-abstract class IValidator {
-  String? validate(String txt);
-}