+++ lib/utils/verify.dart
@@ -1,4 +1,5 @@
-import 'i_validator.dart';
+import 'package:i_validator/validators/i_validator.dart';
+
 
 String? verify({required IValidator iValidator, String? txt}) =>
     iValidator.validate(txt ?? '');