+++ lib/validators/confirm_password_required_validator.dart
@@ -1,0 +1,7 @@
+import 'package:i_validator/i_validator.dart';
+
+class ConfirmPasswordRequiredValidator implements IValidator<String> {
+  @override
+  String? validate(String input) =>
+      input.isEmpty ? kConfirmPasswordMismatchError : null;
+}