+++ lib/validators/i_validator.dart
@@ -1,0 +1,3 @@
+abstract class IValidator<T> {
+  String? validate(T input);
+}